# functions/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "version": "1.0",
        "title": "XYZ Inventory System",
        "description": "XYZ Inventory System Web API Documentation"
    },
    "schemes": [
        "https",
        "http"
    ],
    "servers": [
        {
            "url": "http://localhost:3000/api"
        },
        {
            "url": "http://localhost:3000/api/Dev"
        }
    ],
    "host": "http://localhost:3000",
    "basePath": "/apidocs",
    "tags": [
        {
            "name": "Auth",
            "description": ""
        },
        {
            "name": "User",
            "description": ""
        }
    ],
    "paths": {
        "/user": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Save new user",
                "operationId": "addUser",
                "requestBody": {
                    "description": "Data that need to pass in request body when adding a new user",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "firstName": {
                                        "type": "string",
                                        "description": "first name"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "description": "last lame"
                                    },
                                    "username": {
                                        "type": "string",
                                        "description": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "description": "password"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "User successfully saved",
                        "content": {}
                    },
                    "500": {
                        "description": "Saving user is failed.",
                        "content": {}
                    },
                    "400": {
                        "description": "Invalid ID supplied",
                        "content": {}
                    },
                    "401": {
                        "description": "Authorization information is missing or invalid.",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/product/all": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get all products",
                "description": "Returns list containing all products.",
                "responses": {
                    "200": {
                        "description": "List of Product",
                        "schema": {
                            "type": "array",
                            "items": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "cost": {
                                        "type": "string"
                                    },
                                    "Description": {
                                        "type": "string"
                                    },
                                    "Status": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Product retrieving failed.",
                        "content": {}
                    },
                    "400": {
                        "description": "Invalid ID supplied",
                        "content": {}
                    },
                    "401": {
                        "description": "Authorization information is missing or invalid.",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "",
                "operationId": "login",
                "requestBody": {
                    "description": "Information that need to pass in request body to login.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "description": "username"
                                    },
                                    "password": {
                                        "type": "string",
                                        "description": "password"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Login Success !"
                    },
                    "404": {
                        "description": "Unauthorized access !"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}